requirements.txt
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.0
django-braces==1.12.0
django-crispy-forms==1.7.2
django-disqus==0.5
django-embed-video==1.1.2
django-haystack==2.7.0
django-js-asset==1.1.0
django-localflavor==2.0
django-mptt==0.9.1
django-taggit==0.22.2
django-tinymce==2.7.0
django-tinymce4-lite==1.7.2
idna==2.7
oauthlib==2.0.6
pbr==5.1.1
Pillow==5.0.0
psycopg2==2.7.5
pysolr==3.7.0
python-dateutil==2.6.1
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
social-auth-app-django==2.1.0
social-auth-core==1.7.0

